Time: 4 ms (31.79%) | Memory: 87.7 MB (28.93%) - LeetSync

+++ average-waiting-time/average-waiting-time.java
@@ -13,6 +13,6 @@ class Solution {
             totalWait += customerWait;
             startTime += prepTime;
         }
-        return (double)totalWait * 1.0 / customers.length;
+        return totalWait * 1.0 / customers.length;
     }
 }
